+++ ChangeLog.md
@@ -3,6 +3,13 @@ RDBMS support for the XP Framework: MySQL, Sybase, MSSQL, PostgreSQL, SQLite3, I
 
 ## ?.?.? / ????-??-??
 
+* Fixed SQL cast expression being shadowed by cast function in SQLite3
+  by renaming the latter to `marshal`.
+  (@thekid)
+* Changed rdbms.DriverManager::getConnection() to accept DSN instances
+  as well as strings
+  (@thekid)
+
 ## 6.6.0 / 2016-02-20
 
 * Merged PR #11: Add support for 0x79 tokens - stored procedure return